migrate `ButtonMono`
