+++ packages/app/src/library/Import/Confirm.tsx
@@ -11,7 +11,7 @@ import type { ConfirmProps } from './types';
 import { NotificationsController } from 'controllers/Notifications';
 import { ellipsisFn } from '@w3ux/utils';
 import { ButtonMonoInvert } from 'kits/Buttons/ButtonMonoInvert';
-import { ButtonMono } from 'kits/Buttons/ButtonMono';
+import { ButtonMono } from 'ui-buttons';
 import { useNetwork } from 'contexts/Network';
 
 export const Confirm = ({ address, index, addHandler }: ConfirmProps) => {
