+++ packages/app/src/library/Import/Remove.tsx
@@ -10,7 +10,7 @@ import type { RemoveProps } from './types';
 import { ellipsisFn } from '@w3ux/utils';
 import { NotificationsController } from 'controllers/Notifications';
 import { ButtonMonoInvert } from 'kits/Buttons/ButtonMonoInvert';
-import { ButtonMono } from 'kits/Buttons/ButtonMono';
+import { ButtonMono } from 'ui-buttons';
 import { useNetwork } from 'contexts/Network';
 
 export const Remove = ({ address, getHandler, removeHandler }: RemoveProps) => {
