+++ packages/ui-buttons/src/ButtonMono/index.module.scss
@@ -1,0 +1,9 @@
+// Copyright 2024 @polkadot-cloud/polkadot-staking-dashboard authors & contributors
+// SPDX-License-Identifier: GPL-3.0-only
+
+.btnMono {
+  background: var(--text-color-primary);
+  border: 1px solid var(--text-color-primary);
+  border-radius: var(--button-border-radius-large);
+  color: var(--text-color-invert);
+}
