+++ packages/ui-buttons/src/index.tsx
@@ -3,3 +3,4 @@
 
 export * from './ButtonPrimary';
 export * from './ButtonHelp';
+export * from './ButtonMono';
